requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.2.0
deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docling==2.38.1
docling-core[chunking]==2.38.2
docling-ibm-models==3.6.0
docling-parse==4.1.0
easyocr==1.7.2
et-xmlfile==2.0.0
fastapi==0.115.6
fastapi-utils==0.8.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec[http]==2024.9.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
latex2mathml==3.78.0
lazy-loader==0.4
llama-index==0.12.23
llama-index-agent-openai==0.4.8
llama-index-cli==0.4.1
llama-index-core==0.12.23.post2
llama-index-llms-ollama==0.5.0
llama-index-embeddings-ollama==0.5.0
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.25
llama-index-llms-openai-like==0.3.4
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
markupsafe==2.1.5
marshmallow==3.25.1
mdurl==0.1.2
milvus-lite==2.4.11
milvus-model==0.2.12
mpire[dill]==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.9
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
pyarrow==19.0.0
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.1
pylatexenc==2.10
pymilvus==2.5.4
pypdf==5.1.0
pypdfium2==4.30.1
pytest==8.3.5
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.11.0
redis==5.2.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rtree==1.3.0
safetensors[numpy,torch]==0.5.2
scikit-image==0.25.0
scipy==1.15.1
semchunk==2.2.2
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy[asyncio]==2.0.37
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.1.10
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
wrapt==1.17.2
xlsxwriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
